Remove default template
This change removes the default template.

The default template's purpose is to enable compatibility with origin-only proxy configuration schemes.  This is a worthwhile goal, but the default template is not a suitable solution:
* It violates RFC 8615's definition of .well-known/ as the location for *metadata* about an origin.  A MASQUE service is not metadata.
* It conflicts with better discovery mechanisms, like the one [proposed earlier](https://datatracker.ietf.org/doc/html/draft-schinazi-masque-protocol-03#section-2) in the MASQUE development process.
* It is not compatible with the use of HTTPS records.  (In the targeted legacy context, the client has no other DNS server that can be used when a proxy is in use.)  This is somewhat ironic, as HTTPS records allow bootstrapping QUIC connections, and proxying for QUIC is the primary motivation for CONNECT-UDP.

We should defer this problem to a subsequent draft, and provide a cohesive proxy setup flow, presumably with a standardized configuration object that could be served from .well-known/ and extended to cover CONNECT-IP, DoH and other relevant capabilities.